test: move public API script test out of package suite
